LPD-38542 Move upgrade to head as some customers might skip it due to BPR-7796
