+++ modules/apps/dynamic-data-mapping/dynamic-data-mapping-service/bnd.bnd
@@ -40,6 +40,6 @@ Import-Package:\
 	org.tukaani.*;resolution:=optional,\
 	\
 	*
-Liferay-Require-SchemaVersion: 5.5.2
+Liferay-Require-SchemaVersion: 5.6.1
 Liferay-Service: true
 -dsannotations-options: inherit
