+++ modules/apps/dynamic-data-mapping/dynamic-data-mapping-test/src/testIntegration/java/com/liferay/dynamic/data/mapping/internal/upgrade/v5_6_1/test/DDMFieldAttributeUpgradeProcessTest.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-package com.liferay.dynamic.data.mapping.internal.upgrade.v5_5_2.test;
+package com.liferay.dynamic.data.mapping.internal.upgrade.v5_6_1.test;
 
 import com.liferay.arquillian.extension.junit.bridge.junit.Arquillian;
 import com.liferay.document.library.kernel.model.DLFileEntry;
@@ -121,7 +121,7 @@ public class DDMFieldAttributeUpgradeProcessTest {
 	}
 
 	private static final String _CLASS_NAME =
-		"com.liferay.dynamic.data.mapping.internal.upgrade.v5_5_2." +
+		"com.liferay.dynamic.data.mapping.internal.upgrade.v5_6_1." +
 			"DDMFieldAttributeUpgradeProcess";
 
 	@Inject(
